change token id from u32 to u64
This is done because it is conceivable that we will have more than
4 billion tokens on our chain, particularly if this mechanism is
used to create NFTs or something someday.

There is very little reason from an engineering point of view to
try to save space by using a u32 instead of a u64 -- u32's do
sometimes run out in practice, and a u64 will not here.

+++ api/tests/tokens.rs
@@ -18,5 +18,5 @@ fn test_known_tokens_enum_vs_mc_transaction_core_tokens() {
     }
 
     assert_eq!(known_tokens.len(), 1);
-    assert_eq!(*known_tokens.get("MOB").unwrap() as u32, *tokens::Mob::ID);
+    assert_eq!(*known_tokens.get("MOB").unwrap() as u64, *tokens::Mob::ID);
 }
